requirements.txt
# GeoBotv1 Requirements

# Core Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Optimal Transport
pot>=0.9.0  # Python Optimal Transport

# Causal Inference
networkx>=3.0  # For causal graphs (DAGs)
causalgraphicalmodels>=0.0.4
dowhy>=0.11.0  # For do-calculus and causal inference
pgmpy>=0.1.23  # Probabilistic graphical models

# Bayesian Inference
pymc>=5.0.0  # Modern Bayesian inference
arviz>=0.15.0  # Bayesian visualization and diagnostics
emcee>=3.1.0  # MCMC sampling

# Time Series and State-Space Models
statsmodels>=0.14.0
filterpy>=1.4.5  # Kalman filters
hmmlearn>=0.3.0  # Hidden Markov Models
pykalman>=0.9.5

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
torch>=2.0.0  # For GNNs, VAEs, Transformers
torch-geometric>=2.3.0  # Graph Neural Networks
transformers>=4.30.0  # Hugging Face transformers

# Monte Carlo and Simulation
mesa>=2.0.0  # Agent-based modeling
simpy>=4.0.0  # Discrete event simulation

# Data Ingestion - PDF
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pypdf>=3.15.0
pdfminer.six>=20221105

# Data Ingestion - Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
scrapy>=2.11.0
newspaper3k>=0.2.8  # Article extraction
trafilatura>=1.6.0  # Web scraping and content extraction
selenium>=4.15.0  # For dynamic content
lxml>=4.9.0

# Text Processing and NLP
nltk>=3.8
spacy>=3.6.0
sentence-transformers>=2.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
graphviz>=0.20.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Configuration and CLI
click>=8.1.0
hydra-core>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.82.0

# Jupyter and Development
jupyter>=1.0.0
ipython>=8.14.0
notebook>=7.0.0
